requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
google-genai==1.55.0
requests==2.32.4
pillow==11.3.0
python-dotenv==1.2.1
httpx==0.28.1
